Create new separate module for shared code
First step in moving shared module to own repo

Move validation funcs to shared module
