+++ pkg/artifactory/resource/replication/resource_artifactory_replication_config.go
@@ -10,6 +10,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/artifactory/resource/repository"
 	"github.com/jfrog/terraform-provider-artifactory/v6/pkg/utils"
+	validator "github.com/jfrog/terraform-provider-shared"
 )
 
 type GetReplicationConfig struct {
@@ -32,9 +33,9 @@ var replicationSchemaCommon = map[string]*schema.Schema{
 		Required: true,
 	},
 	"cron_exp": {
-		Type:         schema.TypeString,
-		Required:     true,
-		ValidateFunc: utils.ValidateCron,
+		Type:             schema.TypeString,
+		Required:         true,
+		ValidateDiagFunc: validator.Cron,
 	},
 	"enable_event_replication": {
 		Type:     schema.TypeBool,
@@ -102,8 +103,8 @@ var replicationSchema = map[string]*schema.Schema{
 		Optional: true,
 	},
 	"proxy": {
-		Type:     schema.TypeString,
-		Optional: true,
+		Type:        schema.TypeString,
+		Optional:    true,
 		Description: "Proxy key from Artifactory Proxies setting",
 	},
 }
